# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Vita API Documentation",
        "description": "petshop api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "local server"
        },
        {
            "url": "https://8009.freelancedeveloper.site",
            "description": "remote server"
        }
    ],
    "paths": {
...
        "/api/v1/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get all tasks",
                "operationId": "1adfc7d9e566ea6e76d6344bfb0739f4",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The number of items to display per page (default 15)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tasks found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TaskListSchema"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/components/schemas/PaginatorSchema"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Create a new task",
                "operationId": "30abc4979acc22c17b16032f97e44513",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "block_id",
                                    "goal_id",
                                    "habit_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Task title"
                                    },
                                    "details": {
                                        "type": "string",
                                        "example": "Task details"
                                    },
                                    "block_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "goal_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "habit_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "completed": {
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2022-06-10"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Task created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TaskSchema"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/tasks/{id}/completed": {
            "patch": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Mark task as completed or uncompleted",
                "operationId": "d5eb611889383f606e62404514827132",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the resource",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/tasks/list": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "List tasks",
                "operationId": "c442a1da2eb0a17f5c27fe27b09efe19",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort by day count",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-01-01"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "goal",
                                "habit"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tasks found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tasks": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TaskSchema"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/tasks/{id}/urgent": {
            "patch": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Mark task as completed or uncompleted",
                "operationId": "58c7e1c793c6c5cd6c4702fd4a1cd53f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the resource",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resource updated"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "TaskListSchema": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string"
                    },
                    "details": {
                        "type": "string"
                    },
                    "start": {
                        "type": "string",
                        "example": "2024-01-01 00:00:00"
                    },
                    "end": {
                        "type": "string",
                        "example": "2024-01-01 00:00:00"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}